A code slated for the first vzf beta. After some minor tests and cleaning it will be tagged.

+++ vzfsrc/VZF.Types/Data/Procedures/accessmask_list_Result.cs
@@ -9,14 +9,13 @@
 
 namespace VZF.Types.Data
 {
-    using System;
-    
+
     public partial class accessmask_list_Result
     {
         public int AccessMaskID { get; set; }
         public int BoardID { get; set; }
         public string Name { get; set; }
         public int Flags { get; set; }
-        public short SortOrder { get; set; }
+        public int SortOrder { get; set; }
     }
 }
